move shared image logging into utils
