+++ viscy/translation/engine.py
@@ -23,13 +23,13 @@ from torchmetrics.functional import (
     structural_similarity_index_measure,
 )
 
-from viscy._log_images import detach_sample, render_images
 from viscy.data.typing import Sample
 from viscy.translation.evaluation_metrics import mean_average_precision, ms_ssim_25d
 from viscy.unet.networks.fcmae import FullyConvolutionalMAE
 from viscy.unet.networks.Unet2D import Unet2d
 from viscy.unet.networks.Unet25D import Unet25d
 from viscy.unet.networks.unext2 import UNeXt2
+from viscy.utils.log_images import detach_sample, render_images
 
 try:
     from cellpose.models import CellposeModel
